code: Fix a few 'comparison between signed and unsigned' warnings.

+++ linuxdoom-1.10/src/w_wad.c
@@ -136,7 +136,7 @@ void W_AddFile (char *filename)
 {
     wadinfo_t		header;
     lumpinfo_t*		lump_p;
-    unsigned		i;
+    int			i;
     int			handle;
     int			length;
     int			startlump;
@@ -235,7 +235,7 @@ void W_Reload (void)
     wadinfo_t		header;
     int			lumpcount;
     lumpinfo_t*		lump_p;
-    unsigned		i;
+    int			i;
     int			handle;
     int			length;
     filelump_t*		fileinfo;
@@ -479,7 +479,7 @@ W_CacheLumpNum
 {
     byte*	ptr;
 
-    if ((unsigned)lump >= numlumps)
+    if (lump >= numlumps)
 	I_Error ("W_CacheLumpNum: %i >= numlumps",lump);
 
     if (!lumpcache[lump])
